backend/requirements.txt
Django==5.0.2
djangorestframework==3.15.2
django-filter==24.3
django-environ==0.11.2
google-generativeai
ipython~=8.20.0
Django==5.0.2
djangorestframework==3.15.2
django-filter==24.3
django-environ==0.11.2
django-allauth==0.61.1
dj-rest-auth[with-social]==6.0.0
django-cors-headers==4.4.0
Pillow==10.4.0

